style: heads.py: renamed from get_head_hash.py
